# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /surgeries:
    get:
      operationId: listSurgeries
      tags: [surgery]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Surgery" }
    post:
      operationId: createSurgery
      tags: [surgery]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSurgeryBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Surgery" }
  /surgeries/{id}:
    patch:
      operationId: updateSurgery
      tags: [surgery]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSurgeryBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Surgery" }
...
components:
  schemas:
...
    Surgery:
      type: object
      properties:
        id: { type: integer }
        patientId: { type: integer }
        patientName: { type: string }
        surgeonId: { type: integer }
        surgeonName: { type: string }
        operatingRoom: { type: string }
        scheduledAt: { type: string, format: date-time }
        procedureName: { type: string }
        anesthesia: { type: ["string", "null"] }
        status: { type: string, description: "scheduled|inprogress|completed|cancelled" }
        notes: { type: ["string", "null"] }
      required: [id, patientId, patientName, surgeonId, surgeonName, operatingRoom, scheduledAt, procedureName, status]
    CreateSurgeryBody:
      type: object
      properties:
        patientId: { type: integer }
        surgeonId: { type: integer }
        operatingRoom: { type: string }
        scheduledAt: { type: string, format: date-time }
        procedureName: { type: string }
        anesthesia: { type: ["string", "null"] }
      required: [patientId, surgeonId, operatingRoom, scheduledAt, procedureName]
    UpdateSurgeryBody:
      type: object
      properties:
        status: { type: string }
        notes: { type: ["string", "null"] }
...